p4tc: Initial commit for enabling broadcast POC
Interpret block ID as the pipeline id in the P4 filter.
Whenever a user specifies a block ID, alter the assign the dev group to
the block ID and send a dev set event.

There will be a broadcast module that listens to dev events and will use
the event to add the device to a broadcast list.

Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>