net: sched: act_api: Add support for preallocated dynamic action instances
In P4, actions are assumed to pre exist and have an upper bound number of
instances.

Add the necessary code to preallocate actions instances for dynamic actions.

We add 2 new actions flags:
- TCA_ACT_FLAGS_PREALLOC: Which tells if the actions instance was
  preallocated.
- TCA_ACT_FLAGS_UNUSED: Which tells if the actions instance was
  preallocated but not used yet.

The rule is that preallocated actions can't be deleted by the tc actions
runtime commands and a dump or a get will only show preallocated actions
instances which are being used (TCA_ACT_FLAGS_UNUSED == false).

The preallocated actions will be deleted once the dynamic action kind to
which they belong to is deleted.

Co-developed-by: Victor Nogueira <victor@mojatatu.com>
Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Co-developed-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>