p4tc: add P4 classifier
Introduce P4 tc classifier. A  tc filter instantiated on this classifier
is used to bind a P4 pipeline to one or more netdev ports. To use P4
classifier you must specify a pipeline name that will be associated to
this filter. That pipeline must have already been created via a template.
For example, if we were to add a filter to ingress of network interface
device eth0 and associate it to P4 pipeline simple_l3 we'd issue the
following command:

tc filter add dev lo parent ffff: protocol all prio 6 p4 pname simple_l3 \
    action bpf obj $PROGNAME.o section prog/tc-ingress

Note that we attach a bpf program to the filter in the form of an action.
This will be the representation of the P4 program as generated by the compiler.

The theory of operations is as follows:

================================1 PARSING================================

The parser is implemented in ebpf residing either at the TC or XDP
level.
Note: the different headers and their IDs are already described in the
templating definition and the generated ebpf code.

The parser could be a separate program or integrated into the ebpf program.
For example, when we have a separate parser program which is loaded into
XDP, we need to load it into XDP using tc filter command.

In the filter addition command, we must also specify a path to which the
XDP parser program will be linked to. Below we have an example:

tc filter add dev $P0 ingress protocol all prio 1 p4 pname redirect_srcip \
    prog type xdp obj $PARSER.o section parser/xdp xdp_cookie 22 \
    pinned_link /sys/fs/bpf/mylink \
    action bpf obj $PROGNAME.o section prog/tc-ingress

Note that we also specify a "prog_cookie", which is used to verify
whether the eBPF program has executed or not before we reach the tc P4
classifier. The eBPF program sets this cookie by writing it to the
xdp_md metadata. The TC filter will retrieve it and verify once the
packets reaches it.

=============================ACTIONS=============================

In the above example, the P4 program (minus the parser) is encoded in an action.
It should be noted that classical tc actions continue to work - IOW, someone
could decide to add a mirred action to mirror all packets after or before
the ebpf action.

Co-developed-by: Victor Nogueira <victor@mojatatu.com>
Signed-off-by: Victor Nogueira <victor@mojatatu.com>
Co-developed-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Pedro Tammela <pctammela@mojatatu.com>
Signed-off-by: Jamal Hadi Salim <jhs@mojatatu.com>